+++ gulpfile.js
@@ -4,12 +4,10 @@ var Cesium = require('cesium');
 var child_process = require('child_process');
 var fsExtra = require('fs-extra');
 var gulp = require('gulp');
-var gulpJshint = require('gulp-jshint');
 var Jasmine = require('jasmine');
-var JasmineSpecReporter = require('jasmine-spec-reporter');
+var JasmineSpecReporter = require('jasmine-spec-reporter').SpecReporter;
 var open = require('open');
 var path = require('path');
-var request = require('request');
 var yargs = require('yargs');
 
 var defined = Cesium.defined;
@@ -21,24 +19,7 @@ var environmentSeparator = process.platform === 'win32' ? ';' : ':';
 var nodeBinaries = path.join(__dirname, 'node_modules', '.bin');
 process.env.PATH += environmentSeparator + nodeBinaries;
 
-var jsHintFiles = ['**/*.js', '!node_modules/**', '!coverage/**'];
-var specFiles = ['**/*.js', '!node_modules/**', '!coverage/**'];
-
-gulp.task('jsHint', function () {
-    var stream = gulp.src(jsHintFiles)
-        .pipe(gulpJshint())
-        .pipe(gulpJshint.reporter('jshint-stylish'));
-
-    if (argv.failTaskOnError) {
-        stream = stream.pipe(gulpJshint.reporter('fail'));
-    }
-
-    return stream;
-});
-
-gulp.task('jsHint-watch', function () {
-    gulp.watch(jsHintFiles, ['jsHint']);
-});
+var specFiles = ['**/*.js', '!node_modules/**', '!coverage/**', '!doc/**', '!bin/**'];
 
 gulp.task('test', function (done) {
     var jasmine = new Jasmine();
@@ -54,8 +35,8 @@ gulp.task('test', function (done) {
 
 gulp.task('test-watch', function () {
     gulp.watch(specFiles).on('change', function () {
-        //We can't simply depend on the test task because Jasmine
-        //does not like being run multiple times in the same process.
+        // We can't simply depend on the test task because Jasmine
+        // does not like being run multiple times in the same process.
         try {
             child_process.execSync('jasmine JASMINE_CONFIG_PATH=specs/jasmine.json', {
                 stdio: [process.stdin, process.stdout, process.stderr]
@@ -68,39 +49,14 @@ gulp.task('test-watch', function () {
 
 gulp.task('coverage', function () {
     fsExtra.removeSync('coverage/server');
-    child_process.execSync('istanbul' +
-        ' cover' +
-        ' --include-all-sources' +
+    child_process.execSync('nyc' +
+        ' --all' +
+        ' --reporter=lcov' +
         ' --dir coverage' +
-        ' -x "specs/** coverage/** index.js gulpfile.js"' +
+        ' -x "specs/**" -x "coverage/**" -x "doc/**" -x "bin/**" -x "index.js" -x "gulpfile.js"' +
         ' node_modules/jasmine/bin/jasmine.js' +
         ' JASMINE_CONFIG_PATH=specs/jasmine.json', {
         stdio: [process.stdin, process.stdout, process.stderr]
     });
     open('coverage/lcov-report/index.html');
-});
-
-function copyModule(module) {
-    var tsName = module + '.d.ts';
-    var srcUrl = 'https://raw.githubusercontent.com/DefinitelyTyped/DefinitelyTyped/master/' + module + '/' + tsName;
-    var desPath = path.join('TypeScriptDefinitions', tsName);
-
-    request.get({
-        url: srcUrl
-    }, function (error, response) {
-        if (error) {
-            console.log(error);
-            return;
-        }
-        if (response.statusCode >= 200 && response.statusCode < 300) {
-            fsExtra.outputFileSync(desPath, response.body);
-        }
-    });
-}
-
-gulp.task('update-ts-definitions', function () {
-    fsExtra.removeSync('TypeScriptDefinitions');
-    var packageJson = require('./package.json');
-    Object.keys(packageJson.dependencies).forEach(copyModule);
-    Object.keys(packageJson.devDependencies).forEach(copyModule);
 });