+++ index.js
@@ -1,3 +1,2 @@
-module.exports = {
-    convert : require('./lib/convert')
-};
+'use strict';
+module.exports = require('./lib/obj2gltf');